+++ lib/src/date_picker.dart
@@ -72,11 +72,13 @@ class DatePicker {
       // Check whether date format is legal or not
       if (DateTimeFormatter.isDayFormat(dateFormat)) {
         if (pickerMode == DateTimePickerMode.time) {
-          pickerMode =
-              DateTimeFormatter.isTimeFormat(dateFormat) ? DateTimePickerMode.datetime : DateTimePickerMode.date;
+          pickerMode = DateTimeFormatter.isTimeFormat(dateFormat)
+              ? DateTimePickerMode.datetime
+              : DateTimePickerMode.date;
         }
       } else {
-        if (pickerMode == DateTimePickerMode.date || pickerMode == DateTimePickerMode.datetime) {
+        if (pickerMode == DateTimePickerMode.date ||
+            pickerMode == DateTimePickerMode.datetime) {
           pickerMode = DateTimePickerMode.time;
         }
       }
@@ -99,7 +101,8 @@ class DatePicker {
         onChange: onChange,
         onConfirm: onConfirm,
         theme: Theme.of(context),
-        barrierLabel: MaterialLocalizations.of(context).modalBarrierDismissLabel,
+        barrierLabel:
+            MaterialLocalizations.of(context).modalBarrierDismissLabel,
         minuteDivider: minuteDivider,
       ),
     ).whenComplete(onClose ?? () => {});
@@ -156,12 +159,14 @@ class _DatePickerRoute<T> extends PopupRoute<T> {
   AnimationController createAnimationController() {
     assert(_animationController == null);
     assert(navigator?.overlay != null);
-    _animationController = BottomSheet.createAnimationController(navigator!.overlay!);
+    _animationController =
+        BottomSheet.createAnimationController(navigator!.overlay!);
     return _animationController!;
   }
 
   @override
-  Widget buildPage(BuildContext context, Animation<double> animation, Animation<double> secondaryAnimation) {
+  Widget buildPage(BuildContext context, Animation<double> animation,
+      Animation<double> secondaryAnimation) {
     double height = pickerTheme.pickerHeight;
     if (pickerTheme.title != null || pickerTheme.showTitle) {
       height += pickerTheme.titleHeight;
@@ -204,7 +209,8 @@ class _DatePickerComponent extends StatelessWidget {
           onChange: route.onChange,
           onConfirm: route.onConfirm,
           minuteDivider: route.minuteDivider,
-          onMonthChangeStartWithFirstDate: route.onMonthChangeStartWithFirstDate,
+          onMonthChangeStartWithFirstDate:
+              route.onMonthChangeStartWithFirstDate,
         );
         break;
     }
@@ -214,7 +220,8 @@ class _DatePickerComponent extends StatelessWidget {
         builder: (BuildContext context, Widget? child) {
           return new ClipRect(
             child: new CustomSingleChildLayout(
-              delegate: new _BottomPickerLayout(route.animation!.value, contentHeight: _pickerHeight),
+              delegate: new _BottomPickerLayout(route.animation!.value,
+                  contentHeight: _pickerHeight),
               child: pickerWidget,
             ),
           );