+++ lib/src/date_time_formatter.dart
@@ -31,11 +31,13 @@ class DateTimeFormatter {
   }
 
   /// Split date format to array.
-  static List<String> splitDateFormat(String? dateFormat, {DateTimePickerMode? mode, String? dateFormatSeparator}) {
+  static List<String> splitDateFormat(String? dateFormat,
+      {DateTimePickerMode? mode, String? dateFormatSeparator}) {
     if (dateFormat == null || dateFormat.length == 0) {
       return [];
     }
-    List<String> result = dateFormat.split(RegExp(dateFormatSeparator ?? DATE_FORMAT_SEPARATOR));
+    List<String> result =
+        dateFormat.split(RegExp(dateFormatSeparator ?? DATE_FORMAT_SEPARATOR));
     if (mode == DateTimePickerMode.datetime) {
       // datetime mode need join day format
       List<String> temp = [];
@@ -69,7 +71,8 @@ class DateTimeFormatter {
   }
 
   /// Format datetime string
-  static String formatDateTime(int value, String? format, DateTimePickerLocale locale) {
+  static String formatDateTime(
+      int value, String? format, DateTimePickerLocale locale) {
     if (format == null || format.length == 0) {
       return value.toString();
     }
@@ -109,7 +112,8 @@ class DateTimeFormatter {
   }
 
   /// Format day display
-  static String formatDate(DateTime dateTime, String? format, DateTimePickerLocale locale) {
+  static String formatDate(
+      DateTime dateTime, String? format, DateTimePickerLocale locale) {
     if (format == null || format.length == 0) {
       return dateTime.toString();
     }
@@ -137,19 +141,22 @@ class DateTimeFormatter {
   }
 
   /// format year text
-  static String _formatYear(int value, String format, DateTimePickerLocale locale) {
+  static String _formatYear(
+      int value, String format, DateTimePickerLocale locale) {
     if (format.contains('yyyy')) {
       // yyyy: the digit count of year is 4, e.g. 2019
       return format.replaceAll('yyyy', value.toString());
     } else if (format.contains('yy')) {
       // yy: the digit count of year is 2, e.g. 19
-      return format.replaceAll('yy', value.toString().substring(max(0, value.toString().length - 2)));
+      return format.replaceAll('yy',
+          value.toString().substring(max(0, value.toString().length - 2)));
     }
     return value.toString();
   }
 
   /// format month text
-  static String _formatMonth(int value, String format, DateTimePickerLocale locale) {
+  static String _formatMonth(
+      int value, String format, DateTimePickerLocale locale) {
     List<String>? months;
     if (format.contains('MMMM')) {
       // MMMM: the full name of month, e.g. January
@@ -170,12 +177,14 @@ class DateTimeFormatter {
   }
 
   /// format day text
-  static String _formatDay(int value, String format, DateTimePickerLocale locale) {
+  static String _formatDay(
+      int value, String format, DateTimePickerLocale locale) {
     return _formatNumber(value, format, 'd');
   }
 
   /// format week text
-  static String _formatWeek(int value, String format, DateTimePickerLocale locale) {
+  static String _formatWeek(
+      int value, String format, DateTimePickerLocale locale) {
     if (format.contains('EEEE')) {
       // EEEE: the full name of week, e.g. Monday
       List<String>? weeks = DatePickerI18n.getLocaleWeeks(locale);
@@ -192,17 +201,20 @@ class DateTimeFormatter {
   }
 
   /// format hour text
-  static String _formatHour(int value, String format, DateTimePickerLocale locale) {
+  static String _formatHour(
+      int value, String format, DateTimePickerLocale locale) {
     return _formatNumber(value, format, 'H');
   }
 
   /// format minute text
-  static String _formatMinute(int value, String format, DateTimePickerLocale locale) {
+  static String _formatMinute(
+      int value, String format, DateTimePickerLocale locale) {
     return _formatNumber(value, format, 'm');
   }
 
   /// format second text
-  static String _formatSecond(int value, String format, DateTimePickerLocale locale) {
+  static String _formatSecond(
+      int value, String format, DateTimePickerLocale locale) {
     return _formatNumber(value, format, 's');
   }